requirements.txt
streamlit
ultralytics
opencv-python
pillow
numpy
# streamlit run app.py